requirements.txt
peewee>=2.10.1
python-telegram-bot>=7.0.1
requests>=2.18.4
